Update jenkins-slave.xml
I had this service installed in a location like this:
C:\Program Files\Jenkins
And the service wouldn't start because of the space between Program and Files. I changed locally and the service started without problem. Nevertheless when I started the service from Jenkins, the config file is overwritten, losing this minor fix.